+++ TerrariaServer/Terraria/World/Generation/Searches.cs
@@ -1,8 +1,8 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: Terraria.World.Generation.Searches
-// Assembly: TerrariaServer, Version=1.3.5.1, Culture=neutral, PublicKeyToken=null
-// MVID: C2103E81-0935-4BEA-9E98-4159FC80C2BB
-// Assembly location: F:\Steam\steamapps\common\Terraria\TerrariaServer.exe
+// Assembly: TerrariaServer, Version=1.3.5.3, Culture=neutral, PublicKeyToken=null
+// MVID: 8A63A7A2-328D-424C-BC9D-BF23F93646F7
+// Assembly location: H:\Steam\steamapps\common\Terraria\TerrariaServer.exe
 
 using Microsoft.Xna.Framework;
 
@@ -28,8 +28,8 @@ namespace Terraria.World.Generation
       {
         for (int index = 0; index < this._maxDistance; ++index)
         {
-          if (this.Check(origin.X - index, origin.Y))
-            return new Point(origin.X - index, origin.Y);
+          if (this.Check(origin.X - index, (int) origin.Y))
+            return new Point(origin.X - index, (int) origin.Y);
         }
         return GenSearch.NOT_FOUND;
       }
@@ -48,8 +48,8 @@ namespace Terraria.World.Generation
       {
         for (int index = 0; index < this._maxDistance; ++index)
         {
-          if (this.Check(origin.X + index, origin.Y))
-            return new Point(origin.X + index, origin.Y);
+          if (this.Check(origin.X + index, (int) origin.Y))
+            return new Point(origin.X + index, (int) origin.Y);
         }
         return GenSearch.NOT_FOUND;
       }
@@ -68,8 +68,8 @@ namespace Terraria.World.Generation
       {
         for (int index = 0; index < this._maxDistance; ++index)
         {
-          if (this.Check(origin.X, origin.Y + index))
-            return new Point(origin.X, origin.Y + index);
+          if (this.Check((int) origin.X, origin.Y + index))
+            return new Point((int) origin.X, origin.Y + index);
         }
         return GenSearch.NOT_FOUND;
       }
@@ -88,8 +88,8 @@ namespace Terraria.World.Generation
       {
         for (int index = 0; index < this._maxDistance; ++index)
         {
-          if (this.Check(origin.X, origin.Y - index))
-            return new Point(origin.X, origin.Y - index);
+          if (this.Check((int) origin.X, origin.Y - index))
+            return new Point((int) origin.X, origin.Y - index);
         }
         return GenSearch.NOT_FOUND;
       }
